SNAPSHOT-0.1.1
When you remove a rank, the people who have this rank become "members"
It is now impossible to remove the rank "member"

+++ src/fr/tt54/country/cmd/subcommand/CmdRank.java
@@ -164,6 +164,11 @@ public class CmdRank extends SubCommand {
                         return false;
                     }
 
+                    if (rankToRemove.getName().equalsIgnoreCase("member")) {
+                        player.sendMessage(Main.getMessages().getMessage("cantdeletemember"));
+                        return false;
+                    }
+
                     if (CountryManager.deleteCountryRank(country, name)) {
                         player.sendMessage(Main.getMessages().getMessage("rankdeleted", "%rank%", name));
                         return true;
